new base version
this version is tested and stronger than the latest base version.

- improved learning: uses lower- and upperbound for learning value, automatically determines opponent learning factor and continues learning until the lower/upperbound window is small enough. now has early short-cuts for significant enough results, making the learning process faster. todo: multi-threaded learning
- base piece values for rook and queen adjusted to (465 , 465) and (910, 875)

+++ src/search.h
@@ -42,8 +42,8 @@ enum SEARCH_CONSTANTS {
     HIGH_DEPTH = ONE_PLY * 8
 };
 
-static const int FUTILITY_MARGIN[7] = {
-    2 * VPAWN,
+static const short FUTILITY_MARGIN[7] = {
+    2*VPAWN,
     VKNIGHT, VKNIGHT,
     VROOK, VROOK,
     VQUEEN, VQUEEN
